// imports/source/AplikasiBian-main/frontend/src/lib/types.ts
export type DebtStatus = "unpaid" | "paid";
export type UserRole = "admin" | "kasir";

export interface Product {
  id: string;
  type: ProductType;
  sku: string;
  name: string;
  brand: string;
  size: string;
  stock: number;
  cost_price: number;
  selling_price: number;
  service_fee: number;
}

export interface Customer {
  id: string;
  name: string;
  phone: string;
  address: string;
}

export interface Supplier {
  id: string;
  name: string;
  phone: string;
  address: string;
}
// ...
export interface AccountsReceivable {
  id: string;
  transaction_id: string;
  invoice_number: string;
  customer_id: string;
  customer_name: string;
  amount: number;
  due_date: string;
  status: DebtStatus;
}

export interface AccountsPayable {
  id: string;
  supplier_id: string;
  supplier_name: string;
  invoice_number: string;
  amount: number;
  due_date: string;
  status: DebtStatus;
}